feat: add user guide
